pyproject.toml
[project]
name = "calendar-backend"
version = "0.1.0"
description = "Backend of calendar app"
authors = [
    {name = "donutboyy", email = "dndanielng929@gmail.com"},
]
dependencies = ["fastapi[standard]>=0.115.6", "sqlmodel>=0.0.22", "alembic>=1.14.0", "psycopg2-binary>=2.9.10", "pyjwt>=2.10.1", "passlib[bcrypt]>=1.7.4"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false


[tool.pdm.scripts]
_.env_file = ".env"
start = "fastapi dev app/main.py" 
lint = "ruff check --exclude ./alembic"
format = "black ./app ./tests"

[dependency-groups]
test = [
    "pytest>=8.3.4",
]
dev = [
    "ruff>=0.8.1",
    "black>=24.10.0",
]
